requirements.txt
pandas>=2.0
numpy>=1.24
scipy>=1.11
matplotlib>=3.7
requests>=2.31
pyarrow>=14.0
pyyaml>=6.0
joblib>=1.3
tqdm>=4.65
pytest>=7.0
jsonschema>=4.0
